Assignment - 1st Commit
This is the 3rd Challenge for my iOS8 and Swift online course.

From the instructor:

Objective:
Our goal for this challenge, will be to have the user enter a number in
the textfield, convert it to dog years, and lastly display that new
value in the label.

The key here is to break up the problem into steps. This will help you
when solving this entry level challenge, as well as future bigger
challenges. Feel free to use our recommended steps, or create your own.

Recommended Steps:
- Create a New Xcode Project.
- Implement our Image Assets
- Setup Storyboard Elements
— ImageView
— Textfield
—- Be sure to check which keyboard style you are using with this. We
are only dealing with numbers, so find one that fits that need.
— Label
— Button
- Hook these storyboard elements up with our view controller.
- Figure out how to grab the entered string value from the textfield.
- Convert string variable to an int variable. (*Hint* try using an
optional, but make sure to unwrap it!)
- Create a new variable and set that value to our entered value int * 7
(since the total dog years is equal to human years multiplied by 7).
- Then you want to get that new int and convert it to a string
- Lastly, display this newly converted string to your label.