Extra Work
I added a bit more functionality to this app.

- When the convert button is pressed, it and the input field disappear
and are replaced by the answer and a convert again button.
- When the convert again button is pushed, it and the answer are hidden
and the initial input field and button return.
- There is a fatal crash that occurs in the original program if someone
pushes the convert button and there is nothing entered into the field.
I added a conditional that checks for this and if nothing is inputed it
reports and error to the user and asks for input.
- I also added the ability for the person to enter their dog’s name.