Assignment 2 - Add "Real" Dog Years
In this commit, I do the second assignment for the Age of Laika App for
my iOS8 and Swift class where I need to convert the age given into
“real” dog years; which is 10.5 years for the first 2 years and then 4
years for every year after.

I did this, plus left the initial button for “regular” dog years. The
app still checks for if a name was given for both buttons and displays
a name or “Your dog” if no name given.

Also, I added in the check to see if someone entered “0” as an age, and
if they did, then the app tells them that the dog has not been born
yet.

As this app is now working, I will call it done. But, I could see a few
next steps for added functionality.
- Add a button after a conversion is made for the user to convert the
other age. So, if they push for “Dog Years” then a button to convert
into “Real” dog years would show up with the answer (and vise versa).
- I could also see adding a button on the initial screen to show both
conversions at once.
- I am also sure that there could be some cleaning up of my code to
make it smoother.
- And I could see adding a main screen where the user could select the
type of animal to find the age, and then duplicate the code here to do
an age for a cat, horse, fish, bunny, etc.

But, for now, I will call it done.