Assignment 2 - Clean up
I did clean up a little bit of the code after watching the solution
video from the teacher. I was able to take 3 lines of code and condense
it down to just 1.

Also, I forgot to mention in the previous commit that I also added the
functionality for a user to input a non-whole number for the human age
years. So, 2.5 can be entered, not just 2 or 3. And the app can handle
that double var and do the correct math to it.